% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters}
\usage{
get_parameters(x)
}
\arguments{
\item{x}{an R object of class 'squid'}
}
\description{
Extracts population level data from a squid object
}
