% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{get_population_data}
\alias{get_population_data}
\title{get_population_data}
\usage{
get_population_data(x, list = FALSE, ...)
}
\arguments{
\item{x}{an R object of class 'squid'}

\item{list}{Logical - whether to return data as a list or data_table (FALSE; default).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracts population level data from a squid object
}
