% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_population.R
\name{get_sample_data}
\alias{get_sample_data}
\title{get_sample_data}
\usage{
get_sample_data(x, sample_set = 1, list = FALSE, ...)
}
\arguments{
\item{x}{an R object of class 'squid'}

\item{sample_set}{Integer - which sample set to return. Defaults to 1}

\item{list}{Logical - whether to return data as a list or data_table (FALSE; default).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracts sampled data from a squid object
}
