% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{pop_data}
\alias{pop_data}
\title{pop_data}
\usage{
pop_data(x, ...)
}
\arguments{
\item{x}{an R object of class 'squid'}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracts population level data from a squid object
}
