% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{print.squid}
\alias{print.squid}
\title{print.squid}
\usage{
\method{print}{squid}(x, ...)
}
\arguments{
\item{x}{an R object of class 'squid'}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for class 'squid'
}
