% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_population.R
\name{sample_population}
\alias{sample_population}
\title{sample_population}
\usage{
sample_population(x, type, param, plot = FALSE)
}
\arguments{
\item{x}{A squid object, created using simulate_population().}

\item{type}{Type of sampling, needs to be one of 'nested', 'missing' or temporal. See details.}

\item{param}{A set of parameters, specific to the sampling type. See details.}

\item{plot}{Logical. Should illustrative plots be made - defaults to FALSE.}
}
\value{

}
\description{
Sample population level data
}
\details{
...
}
\examples{

}
