% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_population.R
\name{sim_population}
\alias{sim_population}
\title{sim_population}
\usage{
sim_population(
  parameters,
  data_structure,
  model,
  family = "gaussian",
  link = "identity",
  pedigree
)
}
\arguments{
\item{parameters}{A list of parameters for each hierarchical level. See details.}

\item{data_structure}{A matrix or dataframe with a named column for each grouping factor, inuding the levels}

\item{model}{Optional.}

\item{family}{A description of the error distribution. Default "gaussian".}

\item{link}{A description of the link function distribution. Default "identity".}

\item{pedigree}{A list of pedigrees for each hierarchical level}
}
\value{

}
\description{
Simulate population level data
}
\details{
Parameter list ...
}
\examples{

}
