% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{summary.squid}
\alias{summary.squid}
\title{summary.squid}
\usage{
\method{summary}{squid}(object, ...)
}
\arguments{
\item{object}{an R object of class 'squid'}

\item{...}{further arguments passed to or from other methods.}
}
\description{
summary method for class 'squid'
}
