% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_structure.R
\name{make_structure}
\alias{make_structure}
\title{Make hierarchical structure}
\usage{
make_structure(structure, N, balanced = TRUE)
}
\arguments{
\item{structure}{A formula specifying the structure. See details.}

\item{N}{Number of observations}

\item{balanced}{Logical. Is the number of observations within each grouping level the same? Default TRUE.}
}
\value{

}
\description{
Make hierarchical structure
}
\details{
structure ...
}
\examples{

}
