% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSQUIDfct.R
\name{runSQUIDfct}
\alias{runSQUIDfct}
\title{run SQUID without graphical user interface}
\usage{
runSQUIDfct(input = list(), module = NULL, plot = FALSE,
  X_previsualization = NULL)
}
\arguments{
\item{input}{list of the model input paramters (see details).}

\item{module}{character of the module name. This argument is only used by SQUID app.}

\item{plot}{logical value. If \code{TRUE}, \code{runSQUIDfct} includes the plots of the simulation results into the output list.}

\item{X_previsualization}{character of the environment name. This argument is only used by SQUID app.}
}
\value{
list that includes a data.frame of the full data generated and a data.frame 
                      of the sampled data and plots of the results if \code{plot} is \code{TRUE} (see details).
}
\description{
simulate a world inhabited by individuals whose phenotypes are generated by a user-defined 
                      phenotypic equation and then sampling from this world according to user-defined sampling design.
}
\details{
\tabular{lll}{
 \strong{Parameter name} \tab \strong{Default value} \tab \strong{Description} \cr
 \code{Tmax} \tab 1 \tab Positive integer of the duration of the simulation (number of time steps). \cr
 \code{NP}   \tab 1 \tab Positive integer of the number of replicated populations. 
                         The replicated populations are generated independently 
                         using the same simulation design that has been initially 
                         inputted by the user. Note that data from different 
                         populations will be saved in the same output file that
                         you could later use to run statistical analyses. \cr
 \code{NI}   \tab 1 \tab Positive integer of the number of individuals in each sampled population. 
                         \code{NI} must be divisible by the number of higher-level groups (\code{NG}). \cr
 \code{NT}   \tab 1 \tab Positive integer of the number of traits for each individual. SQUID allows a maximum of 2 traits. \cr
 \code{NG}   \tab 1 \tab Positive integer of the number of higher-level groups. \code{NG} must be lower than the number of individuals. \cr
 
 \code{X1_state}     \tab \code{FALSE} \tab Logical value. If \code{TRUE}, 
                                            an environmental effect x1 will be added to the model equation.  \cr
 \code{X1_sto_state} \tab \code{FALSE} \tab Logical value. If \code{TRUE},
                                            a stochastic environmental effect will be added to the environmental effect x1.
                                            The stochastic environmental effect values have a normal distribution around 0 and a variance of \code{X1_sto_V}\cr
 \code{X1_sto_shared}\tab \code{TRUE}  \tab Logical avlue. If \code{TRUE}, 
                                            the stochastic environmental effect included in the environmental effect x1 
                                            will be shared between individuals. 
                                            If \code{FALSE}, each individual will experience a different stochastic 
                                            environmental effect within the environmental effect x1 
                                            (i.e. a new stochastic environmental effect is generated for each individual).\cr
 \code{X1_sto_V}     \tab 1            \tab Numeric value of the variance used to generate a normal distributed 
                                            stochastic environmental effect within the environmental effect x1.\cr
 
 \code{X1_sto_autocor_state} \tab \code{FALSE} \tab Logical value.  If \code{TRUE}, an autocorrelation effect will be 
                                                    added to the stochastic environemental effect within the environmental effect x1. \cr
 \code{X1_sto_corr}          \tab 0            \tab Numeric value. The correlation value ranging from 0 to 1 
                                                    that characterizes the magnitude of the temporal autocorrelation. \cr
 
 \code{X1_lin_state}     \tab \code{FALSE} \tab Logical value. If \code{TRUE}, a linear trend will be added the environmental effect x1. \cr
 \code{X1_lin_shared}    \tab \code{TRUE}  \tab Logical avlue. If \code{TRUE}, 
                                            the linear trend included in the environmental effect x1 
                                            will be shared between individuals. 
                                            If \code{FALSE}, each individual will experience a different linear trend 
                                            within the environmental effect x1. 
                                            A new linear trend is generated for each individual by varying the intercept 
                                            and the slope of the linear equation fowlling a normal distribution where 
                                            the mean is the intercept and the slope value itself and the variance is \code{X1_lin_V}. \cr
 \code{X1_lin_intercept} \tab 0            \tab Numeric value of the intercept of the linear trend included into the environmental effect x1. \cr
 \code{X1_lin_slope}     \tab 1            \tab Numeric value of the slope of the linear trend included into the environmental effect x1. \cr
 \code{X1_lin_V}         \tab 1            \tab Numeric value of the variance used to generate different (unshared) linear trend effect for each individual within the environmnetal effect x1. \cr
 
 \code{X1_cyc_state}     \tab \code{FALSE} \tab Logical value. If \code{TRUE}, a cyclic trend will be added the environmental effect x1. \cr
 \code{X1_cyc_shared}    \tab \code{TRUE}  \tab Logical avlue. If \code{TRUE}, 
                                                the cyclic trend included in the environmental effect x1 
                                                will be shared between individuals. 
                                                If \code{FALSE}, each individual will experience a different cyclic trend 
                                                within the environmental effect x1.
                                                A new cyclic trend is generated for each individual by varying the amplitude, 
                                                the period, the horizontal shif and the vertical shift 
                                                of the cyclic equation following a normal distribution where 
                                                the mean is the parameter value itself and the variance is \code{X1_cyc_V}. \cr
 \code{X1_cyc_amplitude} \tab 10           \tab Positive numeric value of the amplitude of the cyclic trend included into the environmental effect x1. \cr
 \code{X1_cyc_period}    \tab 10           \tab Positive numeric value of the period of the cyclic trend included into the environmental effect x1. \cr
 \code{X1_cyc_Hshift}    \tab 0            \tab Positive numeric value of the horizontal shift of the cyclic trend included into the environmental effect x1. \cr
 \code{X1_cyc_Vshift}    \tab 0            \tab Positive numeric value of the vertical shift of the cyclic trend included into the environmental effect x1. \cr
 \code{X1_cyc_V}         \tab 0            \tab Numeric value of the variance used to generate different (unshared) cycluc trend effect for each individual within the environmnetal effect x1. \cr
 
 
 \strong{Parameter name} \tab \strong{Default value} \tab \strong{Description}
}
}

