% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSQUIDfct.R
\name{runSQUIDfct}
\alias{runSQUIDfct}
\title{run SQUiD function without graphical user interface}
\usage{
runSQUIDfct(input = list(), module = NULL, plot = FALSE,
  X_previsualization = NULL)
}
\arguments{
\item{input}{list of the model variables.}

\item{module}{character of the module name when used with SQUID shiny app.}
}
\value{
list with a data.frame of the full data generated and a data.frame 
                      of the sampled data.
}
\description{
simulate a world inhabited by individuals whose phenotypes are generated by a user-defined 
                      phenotypic equation and then sampling from this world according to user-defined sampling design.
}

