% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{simulate_population}
\alias{simulate_population}
\title{simulate_population}
\usage{
simulate_population(
  parameters,
  data_structure,
  model,
  family = "gaussian",
  link = "identity",
  pedigree,
  pedigree_type,
  phylogeny,
  phylogeny_type,
  cov_str,
  N,
  N_pop = 1,
  known_predictors,
  extra_betas
)
}
\arguments{
\item{parameters}{A list of parameters for each hierarchical level. See details.}

\item{data_structure}{A matrix or dataframe with a named column for each grouping factor, including the levels}

\item{model}{Optional.}

\item{family}{A description of the error distribution. Default "gaussian".}

\item{link}{A description of the link function distribution. Default "identity".}

\item{pedigree}{A list of pedigrees for each hierarchical level. Each pedigree must be matrix or data.frame, that is at least 3 columns, which correspond to ID, dam and sire.}

\item{pedigree_type}{A list describing what kind of genetic variance is to be simulated from each pedigree. Default is 'additive', other options are 'dominance' and 'epistatic'. Makes use of relationship matrices created by the nadiv package.}

\item{phylogeny}{A list of phylogenies for each hierarchical level. Each pedigree should be phylo class.}

\item{phylogeny_type}{A list describing what mode of evolution should be simulated from each phylogeny. Options are 'brownian'(default) or 'OU'.}

\item{cov_str}{A list of covariance structures for each hierarchical level.}

\item{N}{Sample size when data_structure is not specified}

\item{N_pop}{Number of populations. Default = 1}
}
\value{

}
\description{
Simulate population level data
}
\details{
Parameter list ...
}
\examples{

}
